# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BrandPulse AI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/analyses/estimate": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Cost preview; no side effects",
        "operationId": "estimateAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad_request: malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Start an analysis",
        "operationId": "createAnalysis",
        "parameters": [
          {
            "name": "X-Access-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required only when the run needs new SerpApi calls (LIVE_ACCESS_CODE set).",
              "title": "X-Access-Code"
            },
            "description": "Required only when the run needs new SerpApi calls (LIVE_ACCESS_CODE set)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad_request: malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "live_access_required: new SerpApi calls needed; `X-Access-Code` missing or wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "analysis_not_ready, analysis_failed, signal_not_investigable, investigation_in_progress or live_data_disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "daily_limit_reached, rate_limited (`Retry-After` set) or serpapi_quota_low.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "nlp_unavailable or service_unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Recent analyses",
        "operationId": "listAnalyses",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAnalysesResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Status, stage, progress",
        "operationId": "getAnalysis",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Analysis UUID (or a demo alias when DEMO_MODE=true).",
              "title": "Id"
            },
            "description": "Analysis UUID (or a demo alias when DEMO_MODE=true)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found: unknown or deleted ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{id}/dashboard": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Full dashboard payload",
        "operationId": "getDashboard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Analysis UUID (or a demo alias when DEMO_MODE=true).",
              "title": "Id"
            },
            "description": "Analysis UUID (or a demo alias when DEMO_MODE=true)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found: unknown or deleted ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "analysis_not_ready, analysis_failed, signal_not_investigable, investigation_in_progress or live_data_disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{id}/mentions": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Source items behind an aspect or sentiment",
        "operationId": "listMentions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Analysis UUID (or a demo alias when DEMO_MODE=true).",
              "title": "Id"
            },
            "description": "Analysis UUID (or a demo alias when DEMO_MODE=true)."
          },
          {
            "name": "brand_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default: the target brand.",
              "title": "Brand Id"
            },
            "description": "Default: the target brand."
          },
          {
            "name": "aspect",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "e.g. `battery`.",
              "title": "Aspect"
            },
            "description": "e.g. `battery`."
          },
          {
            "name": "sentiment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Sentiment"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentiment"
            }
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/WindowKind",
              "default": "current"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListMentionsResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found: unknown or deleted ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "top_signal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TopSignal"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "brand",
          "status",
          "created_at"
        ],
        "title": "AnalysisListItem"
      },
      "AnalysisStage": {
        "type": "string",
        "enum": [
          "planning",
          "collecting",
          "processing",
          "analyzing",
          "detecting",
          "snapshotting",
          "done"
        ],
        "title": "AnalysisStage"
      },
      "AnalysisStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "partial",
          "failed"
        ],
        "title": "AnalysisStatus"
      },
      "AnalysisStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "brands": {
            "items": {
              "$ref": "#/components/schemas/BrandRef"
            },
            "type": "array",
            "title": "Brands"
          },
          "product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "serp_calls_used": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Serp Calls Used"
          },
          "serp_calls_budget": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Serp Calls Budget"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "progress",
          "brands",
          "period",
          "serp_calls_used",
          "serp_calls_budget",
          "created_at"
        ],
        "title": "AnalysisStatusResponse",
        "description": "`getAnalysis`: lightweight status, polled every 2 s. Terminal: completed, partial, failed."
      },
...
      "BlockedReason": {
        "type": "string",
        "enum": [
          "serpapi_quota_low",
          "live_data_disabled",
          "daily_limit_reached"
        ],
        "title": "BlockedReason",
        "description": "`estimateAnalysis.blocked_reason` (API.md section 3.11)."
      },
      "BrandRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/BrandRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "BrandRef"
      },
...
      "Category": {
        "type": "string",
        "enum": [
          "consumer_electronics",
          "generic"
        ],
        "title": "Category",
        "description": "Aspect lexicon preset (API.md section 3.1)."
      },
...
      "CreateAnalysisRequest": {
        "properties": {
          "brand": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Brand"
          },
          "product": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "competitors": {
            "items": {
              "type": "string",
              "maxLength": 80,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 2,
            "title": "Competitors"
          },
          "category": {
            "$ref": "#/components/schemas/Category",
            "default": "consumer_electronics"
          },
          "period_days": {
            "type": "integer",
            "enum": [
              7,
              14,
              30
            ],
            "title": "Period Days",
            "default": 30
          },
          "as_of_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          }
        },
        "type": "object",
        "required": [
          "brand"
        ],
        "title": "CreateAnalysisRequest",
        "description": "Body of `createAnalysis` and `estimateAnalysis`."
      },
      "CreateAnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "poll_url": {
            "type": "string",
            "title": "Poll Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "poll_url"
        ],
        "title": "CreateAnalysisResponse"
      },
      "DashboardAnalysis": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "live_run": {
            "type": "boolean",
            "title": "Live Run"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "as_of_date",
          "live_run",
          "period"
        ],
        "title": "DashboardAnalysis"
      },
      "DashboardResponse": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/DashboardAnalysis"
          },
          "target": {
            "$ref": "#/components/schemas/TargetBlock"
          },
          "competitors": {
            "items": {
              "$ref": "#/components/schemas/CompetitorBlock"
            },
            "type": "array",
            "title": "Competitors"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalSummary"
            },
            "type": "array",
            "title": "Signals"
          }
        },
        "type": "object",
        "required": [
          "analysis",
          "target"
        ],
        "title": "DashboardResponse"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "EstimateAnalysisResponse": {
        "properties": {
          "planned_calls": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Planned Calls"
          },
          "cached_calls": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cached Calls"
          },
          "estimated_new_calls": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estimated New Calls"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "serpapi": {
            "$ref": "#/components/schemas/SerpapiQuota"
          },
          "live_enabled": {
            "type": "boolean",
            "title": "Live Enabled"
          },
          "needs_access_code": {
            "type": "boolean",
            "title": "Needs Access Code"
          },
          "can_run": {
            "type": "boolean",
            "title": "Can Run"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockedReason"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "planned_calls",
          "cached_calls",
          "estimated_new_calls",
          "as_of_date",
          "period",
          "serpapi",
          "live_enabled",
          "needs_access_code",
          "can_run"
        ],
        "title": "EstimateAnalysisResponse"
      },
...
      "ListAnalysesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisListItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListAnalysesResponse"
      },
...
      "ListMentionsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MentionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ListMentionsResponse"
      },
...
      "MentionItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "sentiment": {
            "$ref": "#/components/schemas/Sentiment"
          },
          "aspects": {
            "items": {
              "$ref": "#/components/schemas/MentionAspect"
            },
            "type": "array",
            "title": "Aspects"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "sentiment"
        ],
        "title": "MentionItem"
      },
      "Period": {
        "properties": {
          "current_start": {
            "type": "string",
            "format": "date",
            "title": "Current Start"
          },
          "current_end": {
            "type": "string",
            "format": "date",
            "title": "Current End"
          },
          "baseline_start": {
            "type": "string",
            "format": "date",
            "title": "Baseline Start"
          },
          "baseline_end": {
            "type": "string",
            "format": "date",
            "title": "Baseline End"
          }
        },
        "type": "object",
        "required": [
          "current_start",
          "current_end",
          "baseline_start",
          "baseline_end"
        ],
        "title": "Period"
      },
...
      "Sentiment": {
        "type": "string",
        "enum": [
          "positive",
          "neutral",
          "negative"
        ],
        "title": "Sentiment"
      },
...
      "SerpapiQuota": {
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Limit"
          },
          "used": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Used"
          },
          "remaining": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Remaining"
          },
          "reserve": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reserve"
          }
        },
        "type": "object",
        "required": [
          "limit",
          "used",
          "remaining",
          "reserve"
        ],
        "title": "SerpapiQuota"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "web",
          "news",
          "youtube",
          "forum",
          "shopping"
        ],
        "title": "SourceType"
      },
...
      "TopSignal": {
        "properties": {
          "aspect": {
            "type": "string",
            "title": "Aspect"
          },
          "impact": {
            "$ref": "#/components/schemas/ImpactLevel"
          }
        },
        "type": "object",
        "required": [
          "aspect",
          "impact"
        ],
        "title": "TopSignal"
      },
...
      "Warning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisStage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Warning"
      },
      "WindowKind": {
        "type": "string",
        "enum": [
          "baseline",
          "current"
        ],
        "title": "WindowKind"
      }
    }
  }
}